examples moved to a dedicated file
